requirements.txt
# API server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP client (async, used for all external calls)
httpx>=0.27.0

# Data models
pydantic>=2.9.0

# Environment / config
python-dotenv>=1.0.0

# LangChain (Phase II / III)
langchain>=0.3.0
langchain-community>=0.3.0

# YAML (user mapping config)
pyyaml>=6.0.2

# Timezone support
pytz>=2024.1

# Cache client (Redis)
redis>=5.0.0

# Task queue (Phase V)
# celery>=5.4.0
